requirements.txt
pyopengl
pyrr
pillow
numpy
glfw
